Refactor MarkCompleteInput for re-use
We're adding a new task which will have radio buttons to mark the task
as complete. This will need to re-use the MarkCompleteComponent.

Extract the common fields and methods from MarkCompleteInput into a
base class, and add MarkPagesSectionCompleteInput which extends this
base class and has the attributes and methods that apply specifially
to the task to mark the pages section as complete.

A subsequent commit will refactor DeclarationInput which is another
input object that we already pass to the MarkCompleteComponent to
extend the MarkCompleteInput base class.
