+++ config/locales/input_objects/mark_pages_section_complete.yml
@@ -2,7 +2,7 @@ en:
   activemodel:
     errors:
       models:
-        forms/mark_complete_input:
+        forms/mark_pages_section_complete_input:
           attributes:
             base:
               has_routing_errors: This form has validation errors
